bug create bean
