+++ src/main/java/com/virtualclassroom/controller/NewsController.java
@@ -1,9 +1,7 @@
 package com.virtualclassroom.controller;
 
 import com.virtualclassroom.dto.NewsDto;
-import com.virtualclassroom.model.Classroom;
 import com.virtualclassroom.model.News;
-import com.virtualclassroom.model.User;
 import com.virtualclassroom.service.classroom.ClassroomService;
 import com.virtualclassroom.service.news.NewsService;
 import com.virtualclassroom.service.user.UserService;
@@ -13,10 +11,11 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+
+import static java.lang.Long.parseLong;
 
 @Controller
 @RequestMapping("/news")
@@ -24,15 +23,18 @@ public class NewsController {
     private final NewsService newsService;
     private final ClassroomService classroomService;
     private final UserService userService;
-    public NewsController(NewsService newsService, ClassroomService classroomService, UserService userService){this.newsService = newsService;
+
+    public NewsController(NewsService newsService, ClassroomService classroomService, UserService userService) {
+        this.newsService = newsService;
         this.classroomService = classroomService;
         this.userService = userService;
     }
 
+    @PreAuthorize("hasAnyAuthority('TEACHER', 'STUDENT')")
     @GetMapping()
-    public String getNewsPage(@RequestParam String classroomId ,Model model) {
+    public String getNewsPage(@RequestParam Long classroomId,Model model) {
         List<NewsDto> newsDtoList = new ArrayList<>();
-        List<News> newsList = newsService.getNewsByUsername(userService.getCurrentUser().getUserName());
+        var newsList = newsService.getByClassId(classroomId);
         newsList.forEach(news -> {
             var teacherList = userService.findByRoleAndClassroom("TEACHER", news.getId());
             var studentList = userService.findByRoleAndClassroom("STUDENT", news.getId());
