+++ src/main/java/com/virtualclassroom/service/news/NewsService.java
@@ -8,11 +8,7 @@ import java.util.Optional;
 
 public interface NewsService {
     List<News> getAllNews();
-    List<News> getNewsByUsername(String userName);
-
     void addNews(News news);
 
-    Optional<News> findByNewsId(Long id);
-
-    Classroom get(Long id);
+    List<News> getByClassId(Long id);
 }
